requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
pymongo==3.12.0
python-dotenv==0.19.0
gunicorn
redis==4.0.1
pybase62
celery
django-celery-beat



